完成“技能/效果”数据编辑功能 存在无法编辑浮点数的问题，会尽快修正

+++ WarChess/WarChess/scr/dataeditor/horsetablemodule.cpp
@@ -1,6 +1,5 @@
 #include "horsetablemodule.h"
 #include "DataManager.h"
-#include "horsemanager.h"
 
 //#define HORSEMANAGER() DataManager::getSingleton().mHorseManager
 
@@ -149,21 +148,21 @@ bool HorseTableModule::setData(const QModelIndex &index, const QVariant &qvalue,
 		case 3:
 			return DATAMANAGER().SetInt(HORSE_TAG, id, VALUE_TAG, qvalue.toInt());
 		case 4:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_ATTACK, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_ATTACK, qvalue.toFloat());
 		case 5:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_RANGEDATTACK, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_RANGEDATTACK, qvalue.toFloat());
 		case 6:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_DEFENSE, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_DEFENSE, qvalue.toFloat());
 		case 7:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_FORMATION, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_FORMATION, qvalue.toFloat());
 		case 8:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_INITIATIVE, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_INITIATIVE, qvalue.toFloat());
 		case 9:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_ACTIONPOINT, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_ACTIONPOINT, qvalue.toFloat());
 		case 10:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_DETECTION, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_DETECTION, qvalue.toFloat());
 		case 11:
-			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_COVERT, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(HORSE_TAG, id, ATTR_COVERT, qvalue.toFloat());
 		case 12:
 			return DATAMANAGER().SetDataStr(HORSE_TAG, id, SCRIPT_TAG, qvalue.toString().toStdWString());
 		default:
