+++ WarChess/WarChess/scr/dataeditor/soldiertablemodule.cpp
@@ -1,8 +1,5 @@
 #include "soldiertablemodule.h"
 #include "DataManager.h"
-#include "soldiermanager.h"
-
-//#define SOLDIERMANAGER() DataManager::getSingleton().mSoldierManager
 
 const int SoldierTableColumnCount = 16; 
 
@@ -161,25 +158,25 @@ bool SoldierTableModule::setData(const QModelIndex &index, const QVariant &qvalu
 		case 3:
 			return DATAMANAGER().SetInt(SOLDIER_TAG, id, VALUE_TAG, qvalue.toInt());
 		case 4:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_ATTACK, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_ATTACK, qvalue.toFloat());
 		case 5:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_RANGEDATTACK, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_RANGEDATTACK, qvalue.toFloat());
 		case 6:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_DEFENSE, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_DEFENSE, qvalue.toFloat());
 		case 7:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_FORMATION, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_FORMATION, qvalue.toFloat());
 		case 8:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_INITIATIVE, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_INITIATIVE, qvalue.toFloat());
 		case 9:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_ACTIONPOINT, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_ACTIONPOINT, qvalue.toFloat());
 		case 10:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_DETECTION, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_DETECTION, qvalue.toFloat());
 		case 11:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_COVERT, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_COVERT, qvalue.toFloat());
 		case 12:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_INJURY, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_INJURY, qvalue.toFloat());
 		case 13:
-			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_COUNTER, qvalue.toInt());
+			return DATAMANAGER().SetAttribute(SOLDIER_TAG, id, ATTR_COUNTER, qvalue.toFloat());
 		case 14:
 			return DATAMANAGER().SetDataStr(SOLDIER_TAG, id, SCRIPT_TAG, qvalue.toString().toStdWString());
 		case 15:
